+++ replicate.yaml
@@ -1,1 +1,1 @@
-repository: file://data/bert_mlm/.replicate
+repository: file://data//.replicate
